requirements.txt
Backtesting @ git+https://github.com/kernc/backtesting.py.git@0ce24d80b1bcb8120d95d31dc3bb351b1052a27d
beautifulsoup4==4.12.3
bokeh==3.6.2
bottle==0.13.2
cassandra-driver==3.29.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
frozendict==2.4.6
geomet==0.2.1.post1
html5lib==1.1
idna==3.10
Jinja2==3.1.5
lightweight-charts==2.1
lxml==5.3.0
MarkupSafe==3.0.2
multitasking==0.0.11
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pandas_ta==0.3.14b0
peewee==3.17.8
pillow==11.1.0
platformdirs==4.3.6
proxy_tools==0.1.0
pybit==5.8.0
pycryptodome==3.21.0
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-Quartz==11.0
pyobjc-framework-Security==11.0
pyobjc-framework-WebKit==11.0
python-dateutil==2.9.0.post0
pytz==2024.2
pywebview==5.3.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
soupsieve==2.6
ta-lib==0.6.0
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
webencodings==0.5.1
websocket-client==1.8.0
xyzservices==2024.9.0
yfinance==0.2.51
